bài 7,8

+++ src/main/java/vn/com/iviettech/repository/OrderRepository.java
@@ -3,6 +3,7 @@ package vn.com.iviettech.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.iviettech.entity.OrderEntity;
 
-public interface OrderRepository
-        extends JpaRepository<OrderEntity, Integer> {
+public interface OrderRepository extends JpaRepository<OrderEntity, Long> {
+
 }
+
